launcher/CMake: Add symlink to bin on Linux
So it's available in $PATH

Since CMAKE_INSTALL_BINDIR is overridden to CMAKE_INSTALL_LIBDIR on
Linux, hardcode the bin directory as it's universal.

We also have to do a weird dance of a temporary .link file in the build
dir to get this to work correctly for distribution packaging purposes.
See https://github.com/zeek/cmake/blob/master/InstallSymlink.cmake#L24